refactor(auth): reorganize package structure and update import paths #22
