+++ src/main/java/org/ddcn41/ticketing_system/domain/auth/controller/AdminAuthController.java
@@ -1,9 +1,9 @@
-package org.ddcn41.ticketing_system.controller;
+package org.ddcn41.ticketing_system.domain.auth.controller;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
-import org.ddcn41.ticketing_system.dto.AuthDtos.AuthResponse;
-import org.ddcn41.ticketing_system.dto.AuthDtos.LoginRequest;
+import org.ddcn41.ticketing_system.domain.auth.dto.AuthDtos.AuthResponse;
+import org.ddcn41.ticketing_system.domain.auth.dto.AuthDtos.LoginRequest;
 import org.ddcn41.ticketing_system.global.config.JwtUtil;
 import org.ddcn41.ticketing_system.domain.user.entity.User;
 import org.ddcn41.ticketing_system.domain.user.service.UserService;
