+++ src/main/java/org/ddcn41/ticketing_system/domain/auth/controller/PageController.java
@@ -1,4 +1,4 @@
-package org.ddcn41.ticketing_system.controller;
+package org.ddcn41.ticketing_system.domain.auth.controller;
 
 import org.springframework.core.io.ClassPathResource;
 import org.springframework.core.io.Resource;
