+++ src/main/java/org/ddcn41/ticketing_system/domain/auth/dto/response/LogoutResponse.java
@@ -1,10 +1,8 @@
-package org.ddcn41.ticketing_system.dto.response;
+package org.ddcn41.ticketing_system.domain.auth.dto.response;
 
-import lombok.Builder;
 import lombok.Getter;
 
 import lombok.AllArgsConstructor;
-import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
