+++ src/main/java/org/ddcn41/ticketing_system/global/config/JwtAuthenticationFilter.java
@@ -4,7 +4,7 @@ import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import org.ddcn41.ticketing_system.service.TokenBlacklistService;
+import org.ddcn41.ticketing_system.domain.auth.service.TokenBlacklistService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.context.SecurityContextHolder;
